Add  Engine.RunCustom method
By this API the user can pass any net.Listener object to Engine.

It can be used by user's test code which run a Engine and bind to
a free port, then pass the real addres to the test function.

And I think if we added this API then the Engine.RunUnix is
unnecessary.